requirements.txt
logparser3>=1.0.0
pandas>=1.5.0


- drain3 ?
- anonymisation
- zip